+++ LightBulb/Graphics/GraphicObject.hpp
@@ -18,10 +18,10 @@ protected:
 	// A static default font which can be used in all graphic objects
 	static sf::Font font;
 	// Two helper functions to convert coordinates
-	sf::Vector2f calcCartesianFromPolarCoordinates(sf::Vector2f center, double radius, double angle);
-	sf::Vector2f calcCartesianFromPolarCoordinates(sf::Vector2f center, sf::Vector2f radiusVector, double angle);
+    sf::Vector2f calcCartesianFromPolarCoordinates(sf::Vector2f center, float radius, float angle);
+    sf::Vector2f calcCartesianFromPolarCoordinates(sf::Vector2f center, sf::Vector2f radiusVector, float angle);
 	// Calculate the angle of a line
-	double calcAngleFromLine(sf::Vector2f lineStart, sf::Vector2f lineEnd);	
+    float calcAngleFromLine(sf::Vector2f lineStart, sf::Vector2f lineEnd);    
 public:
 	virtual ~GraphicObject() {}
 	GraphicObject();
